requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
grpcio==1.70.0
idna==3.10
jinja2==3.1.5
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.12.0
markdown==3.7
markupsafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pyyaml==6.0.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
soundfile==0.13.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.1
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.12.2
vita==0.0.4
werkzeug==3.1.3
yarl==1.18.3
